Refactors files ahead of 2.3
