+++ src/components.rs
@@ -1,0 +1,19 @@
+use bracket_lib::terminal::RGB;
+use specs::prelude::*;
+use specs_derive::*;
+
+#[derive(Component)]
+pub struct Position {
+    pub x: i32,
+    pub y: i32,
+}
+
+#[derive(Component)]
+pub struct Renderable {
+    pub glyph: bracket_lib::terminal::FontCharType,
+    pub fg: RGB,
+    pub bg: RGB,
+}
+
+#[derive(Component, Debug)]
+pub struct Player {}
